+++ app/protected/folders/[folder_id]/@left_side_bar/page.tsx
@@ -1,5 +1,5 @@
-import { fetchUserFolders } from "@/app/actions";
-import LeftSidebar from "@/components/left-sidebar";
+import LeftSidebar from "@/components/left-sidebar/left-sidebar";
+import { fetchUserFolders } from "@/utils/folders/actions/fetch-user-folders";
 export default async function SidebarLeft() {
   const { content: folders } = await fetchUserFolders();
   return <LeftSidebar initialFolders={folders} />;
