+++ app/protected/study-guides/[study_guide_id]/@left_sidebar/page.tsx
@@ -1,6 +1,5 @@
-import { fetchUserFolders } from "@/app/actions";
-import LeftSidebar from "@/components/left-sidebar";
-
+import LeftSidebar from "@/components/left-sidebar/left-sidebar";
+import { fetchUserFolders } from "@/utils/folders/actions/fetch-user-folders";
 export default async function Page() {
   const { content: folders } = await fetchUserFolders();
   return <LeftSidebar initialFolders={folders} />;
