+++ app/protected/study-guides/[study_guide_id]/layout.tsx
@@ -1,4 +1,4 @@
-import { LandingHeader } from "@/components/landing-header";
+import { LandingHeader } from "@/components/landing-header/landing-header";
 import { SidebarProvider } from "@/components/ui/sidebar";
 import React from "react";
