+++ components/icons/study-guide-icon.tsx
