+++ components/study-guide/operations-study-guide.tsx
